# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Product Passport (DPP) Service API
  description: >-
    REST/HTTPS API for the Digital Product Passport, derived from the
    CEN/CENELEC JTC 24 draft standards:
      * prEN 18222 — APIs for the product passport lifecycle management
        (endpoints, HTTP verbs, parameters, status codes — Tables 1–19)
      * prEN 18223 — System interoperability (payload / semantic data model)
      * prEN 18216 — Data exchange protocols (TLS, JSON/XML/JSON-LD/HTML, error handling)
      * prEN 18239 — Access rights, information security, business confidentiality
      * prEN 18219 — Unique identifiers (incl. W3C DIDs — relevant for DID FlexCo)

    NOTE: The source standards are 2025 drafts (Entwurf / CEN Enquiry) and are
    technology-neutral. This specification is one concrete REST realisation of
    the abstract methods and MUST be re-validated against the final published
    standards. semanticId base URI `https://jtc24/...` is a placeholder from the draft.

    IDENTIFIERS IN PATHS: `dpp_id` and `product_id` are URIs/URLs (prEN 18219)
    and contain `:`, `/` and `.`. They MUST be percent-encoded into a single
    path segment, e.g.
    `https://dpp.example.org/01/09520123456788/8546` becomes
    `https%3A%2F%2Fdpp.example.org%2F01%2F09520123456788%2F8546`.
    The service disables content-type suffix routing (`.json`), so dots inside
    an identifier are preserved.
  version: 0.1.0
  contact:
    name: DPP Service (DID FlexCo)
  license:
...
servers:
...
security:
  - bearerAuth: []   # Public read data is accessible unauthenticated (prEN 18239, §6.2); controlled data requires a token.
paths:
...
  /dpps:
    post:
      tags: [Life Cycle API]
      summary: CreateDPP
      description: Creates a new DPP. Access rules control whether creation is allowed (prEN 18222 §4.6, Table 5).
      operationId: createDPP
      x-semanticId: https://jtc24/dpp/API/CreateDPP/1/0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DigitalProductPassport'
      responses:
        '201':
          description: SuccessCreated — DPP created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigitalProductPassport'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/NotAuthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /dpps/{dppId}:
    parameters:
      - $ref: '#/components/parameters/DppId'
    get:
      tags: [Life Cycle API]
      summary: ReadDPPById
      description: Returns the DPP with the specified DPP ID (prEN 18222 §4.2, Table 1).
      operationId: readDPPById
      x-semanticId: https://jtc24/dpp/API/ReadDPPById/1/0
      responses:
        '200':
          description: Success — requested DPP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DigitalProductPassport' }
            application/ld+json:
              schema: { $ref: '#/components/schemas/DigitalProductPassport' }
            application/xml:
              schema: { $ref: '#/components/schemas/DigitalProductPassport' }
            text/html:
              schema:
                type: string
                description: Human-readable rendering of the DPP (prEN 18216 §5, W3C HTML).
        '401': { $ref: '#/components/responses/NotAuthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      tags: [Life Cycle API]
      summary: UpdateDPP
      description: >-
        Partial update of a DPP, conformant to RFC 7396 (JSON Merge Patch).
        The submitted DPP contains only the data that needs to be updated or added.
        If any part fails, the complete update fails and no changes are adopted.
        All changes are archived per prEN 18221 / Module 6 (prEN 18222 §4.7, Table 6).
      operationId: updateDPPById
      x-semanticId: https://jtc24/dpp/API/UpdateDPPById/1/0
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/DigitalProductPassport'
            example:
              DPPStatus: "Archived"
      responses:
        '200':
          description: Success — updated DPP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DigitalProductPassport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/NotAuthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [Life Cycle API]
      summary: DeleteDPPById
      description: Removes the DPP with the specified DPP ID (end of life). Access rules apply (prEN 18222 §4.8, Table 7).
      operationId: deleteDPPById
      x-semanticId: https://jtc24/dpp/API/DeleteDPPById/1/0
      responses:
        '204':
          description: SuccessNoContent — DPP deleted.
        '401': { $ref: '#/components/responses/NotAuthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /dppsByProductId/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      tags: [Life Cycle API]
      summary: ReadDPPByProductId
      description: Returns the current active DPP (latest version) for the given product ID (prEN 18222 §4.3, Table 2).
      operationId: readDPPByProductId
      x-semanticId: https://jtc24/dpp/API/ReadDPPByProductId/1/0
      responses:
        '200':
          description: Success — latest active DPP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DigitalProductPassport' }
        '401': { $ref: '#/components/responses/NotAuthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /dppsByProductIdAndDate/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
      - name: date
        in: query
        required: true
        description: UTC timestamp for which the DPP version is requested.
        schema:
          type: string
          format: date-time
    get:
      tags: [Life Cycle API]
      summary: ReadDPPVersionByProductIdAndDate
      description: Returns the DPP version valid at the given date (prEN 18222 §4.4, Table 3). Optional for the economic operator.
      operationId: readDPPVersionByProductIdAndDate
      x-semanticId: https://jtc24/dpp/API/ReadDPPVersionByProductIdAndDate/1/0
      responses:
        '200':
          description: Success — DPP version for the given date.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DigitalProductPassport' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /dppsByProductIds:
    post:
      tags: [Life Cycle API]
      summary: ReadDPPIdsByProductIds
      description: >-
        Returns a list of DPP identifiers matching a set of product identifiers
        (AND-match of the specified product labels). Supports cursor pagination
        (prEN 18222 §4.5, Table 4).
      operationId: readDPPIdsByProductIds
      x-semanticId: https://jtc24/dpp/API/ReadDPPIdsByProductIds/1/0
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum size of the result set.
          schema: { type: integer, minimum: 0 }
        - name: cursor
          in: query
          required: false
          description: Position from which to resume a result listing (shall not be empty).
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              description: Set of product identifiers (key-value product-identifying information).
              items: { $ref: '#/components/schemas/Identifier' }
      responses:
        '200':
          description: Success — matching DPP identifiers.
          content:
            application/json:
              schema:
                type: object
                required: [payload]
                properties:
                  payload:
                    type: array
                    items: { $ref: '#/components/schemas/Identifier' }
                  nextCursor:
                    type: string
                    nullable: true
                    description: Cursor to fetch the next page, or null if no further results.
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InternalError' }

  ###########################################################################
  # 5  Registry API for Register
  ###########################################################################
...
  /dpps/{dppId}/elements/{elementPath}:
    parameters:
      - $ref: '#/components/parameters/DppId'
      - name: elementPath
        in: path
        required: true
        description: Absolute ElementId path to the specific data element within the DPP.
        schema: { type: string }
    get:
      tags: [Fine Granular Life Cycle API]
      summary: ReadDataElement
      description: Returns a specific data element by its absolute path (prEN 18222 §6.3, Table 10).
      operationId: readDataElement
      x-semanticId: https://jtc24/dpp/API/ReadDataElement/1/0
      responses:
        '200':
          description: Success — requested data element.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataElement' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Fine Granular Life Cycle API]
      summary: UpdateDataElement
      description: >-
        Updates, enhances or removes a specific data element within a DPP
        (prEN 18222 §6.5, Table 12).
      operationId: updateDataElement
      x-semanticId: https://jtc24/dpp/API/UpdateDataElement/1/0
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/DataElement' }
      responses:
        '200':
          description: Success — updated data element.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DataElement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OAuth 2.0 / OpenID Connect access token (JWT). Public data is readable
        without authentication; controlled data requires an authenticated actor
        with an appropriate role (authority, refurbisher, consumer, ...) per prEN 18239.
  parameters:
    DppId:
      name: dppId
      in: path
      required: true
      description: The DPP's globally unique identifier.
      schema: { $ref: '#/components/schemas/Identifier' }
    ProductId:
      name: productId
      in: path
      required: true
      description: The product identifier (enables a web link to the DPP).
      schema: { $ref: '#/components/schemas/Identifier' }

  ###########################################################################
  # Reusable error responses (prEN 18222 §7, Table 16)
  ###########################################################################
  responses:
    BadRequest:
      description: ClientErrorBadRequest — bad or malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
    NotAuthorized:
      description: ClientNotAuthorized — wrong or missing authorization credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
    Forbidden:
      description: ClientForbidden — authorization refused.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
    NotFound:
      description: ClientErrorResourceNotFound — resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
    Conflict:
      description: ClientResourceConflict — resource already exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
    InternalError:
      description: ServerInternalError — unexpected error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
...
  schemas:
    Identifier:
      type: string
      description: Globally unique identifier; SHOULD be a URI/URL (see prEN 18219, incl. W3C DID schemes).
      example: "https://dpp.example.org/01/09520123456788/21/12345"

    # ---- prEN 18222 §7: Status / Result / Message ------------------------
    StatusCode:
      type: string
      description: Generic status code of the response (prEN 18222 Table 16).
      enum:
        - Success
        - SuccessCreated
        - SuccessAccepted
        - SuccessNoContent
        - ClientErrorBadRequest
        - ClientNotAuthorized
        - ClientForbidden
        - ClientMethodNotAllowed
        - ClientErrorResourceNotFound
        - ClientResourceConflict
        - ServerInternalError
        - ServerErrorBadGateway
    MessageTypeEnum:
      type: string
      enum: [Info, Warning, Error, Exception]
      description: The message type (prEN 18222 Table 15).
    Message:
      type: object
      description: A message with more information for the requester (prEN 18222 Table 14).
      required: [messageType, text]
      properties:
        messageType: { $ref: '#/components/schemas/MessageTypeEnum' }
        text: { type: string }
        code:
          type: string
          description: Technology-dependent status or error code.
        correlationId:
          type: string
          description: Identifier to relate several result messages across systems.
        timestamp:
          type: string
          format: date-time
    Result:
      type: object
      description: Result object returned on failed method execution (prEN 18222 Table 13).
      properties:
        statusCode: { $ref: '#/components/schemas/StatusCode' }
        message:
          type: array
          items: { $ref: '#/components/schemas/Message' }

    # ---- prEN 18223 §4.1: DPP semantic model -----------------------------
    DigitalProductPassport:
      type: object
      description: A Digital Product Passport instance (prEN 18223 Table 1 + §4.1.3.1).
      required:
        - DigitalProductPassportID
        - ProductID
        - Granularity
        - DPPSchemaVersion
        - DPPStatus
        - LastUpdate
        - EconomicOperatorID
      properties:
        DigitalProductPassportID:
          allOf: [{ $ref: '#/components/schemas/Identifier' }]
          description: >-
            Unique, globally unique identifier of the DPP instance (URI/URL).
            On CreateDPP this MAY be omitted: the service then mints a did:oyd
            (prEN 18219) and returns it as the dpp ID (Variante A).
        ProductID:
          allOf: [{ $ref: '#/components/schemas/Identifier' }]
          description: Unique product identifier that enables a web link to the DPP.
        Granularity:
          type: string
          description: Level of granularity of the ProductID as per ESPR.
          enum: [model, batch, item]
        DPPSchemaVersion:
          type: string
          description: Reference standard/schema version the DPP instance refers to.
        DPPStatus:
          type: string
          description: Status of the DPP instance as a digital resource.
          enum: [Active, Archived]
        LastUpdate:
          type: string
          format: date-time
          description: Date and time (UTC) of the latest update to the DPP instance.
        EconomicOperatorID:
          allOf: [{ $ref: '#/components/schemas/Identifier' }]
          description: Identifier of an actor involved in the product's life cycle.
        FacilityID:
          allOf: [{ $ref: '#/components/schemas/Identifier' }]
          description: Identifier of a location/building in the value chain (optional).
        dataElementCollections:
          type: array
          items: { $ref: '#/components/schemas/DataElementCollection' }
        dataElements:
          type: array
          items: { $ref: '#/components/schemas/DataElement' }
    DataElementCollection:
      type: object
      description: Collection of DataElements (prEN 18223 §4.1.3.2, Table 2).
      required: [ElementId, DataElements]
      properties:
        ElementId:
          type: string
          description: Relative identifier, unique within its location; absolute path unique within the DPP.
        DictionaryReference:
          type: string
          description: Unique identifier of the semantic description of the collection.
        Name:
          type: string
          description: Human-readable name (translatable).
        DataElements:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/DataElement' }
    DataElement:
      description: >-
        A single data point (prEN 18223 §4.1.3.3). Abstract — realised as a
        SinglevaluedDataElement or MultivaluedDataElement, discriminated by `@type`.
      oneOf:
        - $ref: '#/components/schemas/SinglevaluedDataElement'
        - $ref: '#/components/schemas/MultivaluedDataElement'
      discriminator:
        propertyName: '@type'
        mapping:
          SinglevaluedDataElement: '#/components/schemas/SinglevaluedDataElement'
          MultivaluedDataElement: '#/components/schemas/MultivaluedDataElement'
    SinglevaluedDataElement:
      type: object
      description: A single data point with one value (prEN 18223 §4.1.3.4, Table 3).
      required: ['@type', ElementId, Value]
      properties:
        '@type':
          type: string
          enum: [SinglevaluedDataElement]
        ElementId: { type: string }
        DictionaryReference: { type: string }
        Name: { type: string }
        Value:
          description: The data value (any JSON datatype).
          nullable: true
        ValueDataType:
          type: string
          description: >-
            XSD data type of the value (W3C XML Schema), excluding list types
            (ENTITIES, IDREFS, NMTOKENS), string built-ins and NOTATION/QName.
        Language:
          type: string
          description: BCP 47 language tag (e.g. "en-GB"); ignored for translatable strings.
        UnitOfMeasure:
          type: string
          description: >-
            Optional unit of the value — only for a Product Property
            (prEN 18223 §4.1.4.1, Table 6), e.g. "lm", "W", "K", "h".
          example: lm
        Tolerance:
          type: string
          description: Optional tolerance of the value (Product Property, prEN 18223 Table 6).
    MultivaluedDataElement:
      type: object
      description: A single data point containing a list of values (prEN 18223 §4.1.3.5, Table 4).
      required: ['@type', ElementId, ValueList]
      properties:
        '@type':
          type: string
          enum: [MultivaluedDataElement]
        ElementId: { type: string }
        DictionaryReference: { type: string }
        Name: { type: string }
        ValueList:
          type: array
          items: { $ref: '#/components/schemas/ValueElement' }
    ValueElement:
      type: object
      description: A value associated to a MultivaluedDataElement (prEN 18223 §4.1.3.6, Table 5).
      required: [ElementId, Value, ValueDataType]
      properties:
        ElementId: { type: string }
        DictionaryReference: { type: string }
        Name: { type: string }
        Value:
          description: The data value (any JSON datatype).
          nullable: true
        ValueDataType: { type: string }
        Language: { type: string }
        UnitOfMeasure:
          type: string
          description: Optional unit (Product Property, prEN 18223 Table 6).
        Tolerance:
          type: string
          description: Optional tolerance (Product Property, prEN 18223 Table 6).

    # ---- prEN 18222 §5.2: Registry registration payload (DPP header subset)
...